Template imput field to select field
Transform the template imput field, in a select field, filled with the name of all the template files you find in a [child]theme "flexible-posts-widget" folder, or alternatively in the plugin "views" folder.
I think that a dropdown like the one you have, for example, when it comes to select a page template, could be a better choice. The developer can create as much templates as needed and give them a meaningful file name, and the customer has just to select the appropriate one from a precompiled list.